frontend: Refactor housing company directory file paths
