+++ frontend/src/features/housingCompany/HousingCompanyDetailsPage.tsx
@@ -18,7 +18,7 @@ import {IHousingCompanyDetails, IPropertyManager} from "../../common/schemas";
 import {formatAddress, formatDate, formatMoney} from "../../common/utils";
 import {HousingCompanyApartmentResultsList} from "../apartment/ApartmentListPage";
 import {BatchCompleteApartmentsModal} from "./";
-import SalesCatalogImport from "./components/housingCompany/SalesCatalogImport";
+import SalesCatalogImport from "./components/SalesCatalogImport";
 
 const LoadedHousingCompanyDetails = ({housingCompany}: {housingCompany: IHousingCompanyDetails}): React.JSX.Element => {
     const [isModifyPropertyManagerModalVisible, setIsModifyPropertyManagerModalVisible] = useState(false);
